Broke output flag into seperate command template
