+++ src/main/resources/org/openstreetmap/atlas/utilities/command/subcommands/templates/AtlasLoaderCommandSection.txt
@@ -7,16 +7,15 @@ examples:
 #$ example-loader-command atlas1.atlas atlas2.atlas atlas3.atlas --strict
 If an atlas loader command cannot find one of the specified atlas files, it will
 simply skip over loading the file. This is unless the '--strict' option is applied,
-in which case the load will fail fast. Atlas loader commands also provide an '--output' option,
-which can be used to change the default location of any output files. This example changes
-the output location of an example command to your Desktop:
-#$ example-loader-command myatlas1.atlas --output ~/Desktop
+in which case the load will fail fast.
+
 The '--combine' option forces the atlas loader command to process their input atlases as a single MultiAtlas.
 The MultiAtlas construction step will occur before the process step, and then the processing will be applied to
 the newly combined MultiAtlas. The combined MultiAtlas will always have the name 'combined.atlas'.
 
 The '--parallel' option forces the atlas loader command to process the input atlases in parallel.
 
+Note that any atlas loader command automatically inherits multiple output behaviour.
 In this final example, we strictly process all atlases in your home folder in parallel, and output the results
 to your desktop.
 #$ example-loader-command --strict ~/*.atlas --parallel --output=/Users/$(whoami)/Desktop
